requirement.txt
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
cycler==0.11.0
distlib==0.3.6
dnspython==2.2.1
email-validator==1.2.1
filelock==3.8.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.37.1
greenlet==1.1.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
mailjet-rest==1.3.4
MarkupSafe==2.1.1
matplotlib==3.5.3
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.23.2
packaging==21.3
passlib==1.7.4
Pillow==9.2.0
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.5.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.40
urllib3==1.26.12
virtualenv==20.16.4
Werkzeug==2.2.2
WTForms==3.0.1
